Fix `super-with-arguments` if no `__class__` cell
In certain cases, it's not actually possible to use Python 3's `super()`
without any arguments. A simple example would be a function that is
meant to decorate a class and therefore does not have a known
`__class__` cell. This decorator must instead be explicitly given a
`cls`.

This fixes an error with the `super-with-arguments` check, introduced in
version 2.6.0.

The root cause is that the return type of `node_frame_class()` is
`Optional`, and we don't check for the null case.

After searching recent pylint issues, it does appear that somebody else
already noticed & applied the same fix. See:
https://github.com/PyCQA/pylint/pull/3810

This includes a small test case, though. The original author should
receive credit, but hopefully this is helpful.

------------------------------------------------------------------------

This test passes with:
```
python3 -m pytest "tests/test_functional.py::test_functional[super_with_arguments]"
```

(the test also fails without this change)